# spec/openapi.yaml
openapi: 3.0.3
info:
  title: Grid Stream API Spec - OpenAPI 3.0
  version: 1.0.12
  description: |
    API spec for Grid Stream with updated schema.
  contact:
    email: eric.cuenat@unb.ca
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: TODO
paths:
...
  /v1/der-metadata:
    post:
      tags:
        - der-metadata
      summary: Create DER metadata
      description: >
        Creates metadata for a Distributed Energy Resource (DER). 
        The `id` will be generated by the backend and included in the response.
      operationId: createDERMetadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DERMetadata'
      responses:
        '201':
          description: Successfully created DER metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "DERMetaID_123456789"
                  derMetadata:
                    $ref: '#/components/schemas/DERMetadata'
        '400':
          description: Invalid data provided
        '401':
          description: Unauthorized request
      security:
        - firebase_auth: []
  /v1/der-metadata/{id}:
    get:
      tags:
        - der-metadata
      summary: Get DER metadata by ID
      description: Returns metadata for a Distributed Energy Resource (DER) by its ID
      operationId: getDERMetadataById
      parameters:
        - name: id
          in: path
          description: ID of the DER to retrieve metadata for
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully found DER metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DERMetadata'
        '404':
          description: DER metadata not found
        '400':
          description: Invalid DER ID
        '401':
          description: Unauthorized request from user
      security:
        - firebase_auth: []
    put:
      tags:
        - der-metadata
      summary: Update existing DER metadata
      description: Updates the metadata for a Distributed Energy Resource (DER)
      operationId: updateDERMetadata
      parameters:
        - name: id
          in: path
          description: ID of the DER to update metadata for
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DERMetadata'
      responses:
        '200':
          description: Successfully updated DER metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DERMetadata'
        '404':
          description: DER metadata not found
        '400':
          description: Invalid DER data
        '401':
          description: Unauthorized request from user
      security:
        - firebase_auth: []
    delete:
      tags:
        - der-metadata
      summary: Delete DER metadata
      description: Deletes metadata for a Distributed Energy Resource (DER) by its ID
      operationId: deleteDERMetadata
      parameters:
        - name: id
          in: path
          description: ID of the DER to delete metadata for
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successfully deleted DER metadata
        '404':
          description: DER metadata not found
        '400':
          description: Invalid DER ID
        '401':
          description: Unauthorized request from user
      security:
        - firebase_auth: []
...
components:
  schemas:
...
    DERMetadata:
      type: object
      properties:
        id:
          type: string
        project_id:
          type: string
        type:
          type: string
        nameplate_capacity:
          type: number
          format: float
        power_capacity:
          type: number
          format: float
...
  securitySchemes:
    firebase_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Use Firebase ID token as the Bearer token.
    service_account_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Authentication using a GCP service account token. 
    api_key:
      type: apiKey
      name: api_key
      in: header